Swap dark bits of background to fix sidebar
Swap the dark bits to a light grey to allow us to read the light text in the left hand sidebar and in the footer.

Quick fix to make the style less obviously broken. We need to investigate why the skin has reverted to looking like vector, with all our nice customisations overriden. The dark background parts are the only surviving customisations actually (but not working well with the text)